+++ apps/condo/domains/acquiring/access/PaymentRule.js
@@ -1,6 +1,6 @@
 // @ts-nocheck
 /**
- * Generated by `createschema acquiring.PaymentRule 'context:Relationship:AcquiringIntegrationContext:PROTECT;explicitFee?:Decimal;explicitServiceCharge?:Decimal;implicitFee?:Decimal;bankAccount?:Relationship:BankAccount:PROTECT;marketPlaceScope?:Relationship:PaymentRuleMarketPlaceScope:PROTECT;billingScope?:Relationship:PaymentRuleBillingScope:PROTECT'`
+ * Generated by `createschema acquiring.PaymentRule 'context:Relationship:AcquiringIntegrationContext:PROTECT;explicitFee?:Decimal;explicitServiceCharge?:Decimal;implicitFee?:Decimal;bankAccount?:Relationship:BankAccount:PROTECT;marketPlaceScope?:Relationship:PaymentRuleMarketScope:PROTECT;billingScope?:Relationship:PaymentRuleBillingScope:PROTECT'`
  */
 
 const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
