--- apps/condo/domains/acquiring/access/PaymentRuleMarketPlaceScope.js
@@ -1,29 +1,0 @@
-// @ts-nocheck
-/**
- * Generated by `createschema acquiring.PaymentRuleMarketPlaceScope 'paymentRule:Relationship:PaymentRule:PROTECT;property?:Relationship:Property:PROTECT;squIds?:Json'`
- */
-
-const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
-
-async function canReadPaymentRuleMarketPlaceScopes ({ authentication: { item: user } }) {
-    if (!user) return throwAuthenticationError()
-    if (user.deletedAt) return false
-
-    return user.isAdmin || user.isSupport
-}
-
-async function canManagePaymentRuleMarketPlaceScopes ({ authentication: { item: user }, originalInput, operation, itemId }) {
-    if (!user) return throwAuthenticationError()
-    if (user.deletedAt) return false
-
-    return user.isAdmin || user.isSupport
-}
-
-/*
-  Rules are logical functions that used for list access, and may return a boolean (meaning
-  all or no items are available) or a set of filters that limit the available items.
-*/
-module.exports = {
-    canReadPaymentRuleMarketPlaceScopes,
-    canManagePaymentRuleMarketPlaceScopes,
-}
