+++ apps/condo/domains/acquiring/access/PaymentRuleMarketScope.js
@@ -1,0 +1,29 @@
+// @ts-nocheck
+/**
+ * Generated by `createschema acquiring.PaymentRuleMarketScope 'paymentRule:Relationship:PaymentRule:PROTECT;property?:Relationship:Property:PROTECT;squIds?:Json'`
+ */
+
+const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
+
+async function canReadPaymentRuleMarketScopes ({ authentication: { item: user } }) {
+    if (!user) return throwAuthenticationError()
+    if (user.deletedAt) return false
+
+    return user.isAdmin || user.isSupport
+}
+
+async function canManagePaymentRuleMarketScopes ({ authentication: { item: user }, originalInput, operation, itemId }) {
+    if (!user) return throwAuthenticationError()
+    if (user.deletedAt) return false
+
+    return user.isAdmin || user.isSupport
+}
+
+/*
+  Rules are logical functions that used for list access, and may return a boolean (meaning
+  all or no items are available) or a set of filters that limit the available items.
+*/
+module.exports = {
+    canReadPaymentRuleMarketScopes,
+    canManagePaymentRuleMarketScopes,
+}
