+++ apps/condo/domains/acquiring/constants/errors.js
@@ -84,9 +84,9 @@ const PAYMENT_RULE_BILLING_SCOPE_INVALID_SERVICE_IDS = 'PAYMENT_RULE_BILLING_SCO
 const PAYMENT_RULE_BILLING_SCOPE_MISSING_ALL_FIELDS = '[paymentRuleBillingScope:emptyFields] At least one field besides payment rule must be present'
 const PAYMENT_RULE_BILLING_SCOPE_PROPERTY_AND_PAYMENT_RULE_ORGANIZATIONS_DOES_NOT_MATCH = '[paymentRuleBillingScope:property:organization:noMatch] Property must be of same organization, as payment rule'
 
-const PAYMENT_RULE_MARKETPLACE_SCOPE_INVALID_SKU_IDS = 'PAYMENT_RULE_MARKETPLACE_SCOPE_INVALID_SKU_IDS'
-const PAYMENT_RULE_MARKETPLACE_SCOPE_MISSING_ALL_FIELDS = '[paymentRuleMarketScope:emptyFields] At least one field besides payment rule must be present'
-const PAYMENT_RULE_MARKETPLACE_SCOPE_PROPERTY_AND_PAYMENT_RULE_ORGANIZATIONS_DOES_NOT_MATCH = '[paymentRuleMarketScope:property:organization:noMatch] Property must be of same organization, as payment rule'
+const PAYMENT_RULE_MARKET_SCOPE_INVALID_SKU_IDS = 'PAYMENT_RULE_MARKET_SCOPE_INVALID_SKU_IDS'
+const PAYMENT_RULE_MARKET_SCOPE_MISSING_ALL_FIELDS = '[paymentRuleMarketScope:emptyFields] At least one field besides payment rule must be present'
+const PAYMENT_RULE_MARKET_SCOPE_PROPERTY_AND_PAYMENT_RULE_ORGANIZATIONS_DOES_NOT_MATCH = '[paymentRuleMarketScope:property:organization:noMatch] Property must be of same organization, as payment rule'
 
 const RECURRENT_PAYMENT_CONTEXT_BOTH_TRIGGER_SET_UP_ERROR = 'RECURRENT_PAYMENT_CONTEXT_BOTH_TRIGGER_SET_UP_ERROR'
 const RECURRENT_PAYMENT_CONTEXT_NO_TRIGGER_SET_UP_ERROR = 'RECURRENT_PAYMENT_CONTEXT_NO_TRIGGER_SET_UP_ERROR'
@@ -210,9 +210,9 @@ module.exports = {
     PAYMENT_RULE_NOT_EXISTS,
     PAYMENT_RULE_BILLING_SCOPE_MISSING_ALL_FIELDS,
     PAYMENT_RULE_BILLING_SCOPE_PROPERTY_AND_PAYMENT_RULE_ORGANIZATIONS_DOES_NOT_MATCH,
-    PAYMENT_RULE_MARKETPLACE_SCOPE_INVALID_SKU_IDS,
-    PAYMENT_RULE_MARKETPLACE_SCOPE_MISSING_ALL_FIELDS,
-    PAYMENT_RULE_MARKETPLACE_SCOPE_PROPERTY_AND_PAYMENT_RULE_ORGANIZATIONS_DOES_NOT_MATCH,
+    PAYMENT_RULE_MARKET_SCOPE_INVALID_SKU_IDS,
+    PAYMENT_RULE_MARKET_SCOPE_MISSING_ALL_FIELDS,
+    PAYMENT_RULE_MARKET_SCOPE_PROPERTY_AND_PAYMENT_RULE_ORGANIZATIONS_DOES_NOT_MATCH,
     RECURRENT_PAYMENT_CONTEXT_BOTH_TRIGGER_SET_UP_ERROR,
     RECURRENT_PAYMENT_CONTEXT_NO_TRIGGER_SET_UP_ERROR,
     RECURRENT_PAYMENT_CONTEXT_PAYMENT_DAY_WRONG_RANGE_ERROR,
