+++ apps/condo/domains/acquiring/schema/PaymentRuleBillingScope.test.js
@@ -2,24 +2,24 @@
  * Generated by `createschema acquiring.PaymentRuleBillingScope 'paymentRule:Relationship:PaymentRule:PROTECT;property?:Relationship:Property:PROTECT;category?:Relationship:BillingCategory:PROTECT;bankAccountNumber?:Text;serviceId?:Json'`
  */
 
-const { makeLoggedInAdminClient, makeClient, UUID_RE, DATETIME_RE, waitFor, expectValuesOfCommonFields,
-    expectToThrowValidationFailureError, expectToThrowGQLError
+const { makeLoggedInAdminClient, makeClient, UUID_RE, DATETIME_RE, expectValuesOfCommonFields,
+    expectToThrowValidationFailureError, expectToThrowGQLError,
 } = require('@open-condo/keystone/test.utils')
-
 const {
     expectToThrowAuthenticationErrorToObj, expectToThrowAuthenticationErrorToObjects,
     expectToThrowAccessDeniedErrorToObj, expectToThrowAccessDeniedErrorToObjects,
 } = require('@open-condo/keystone/test.utils')
 
+const { PaymentRuleBillingScope, createTestPaymentRuleBillingScope, updateTestPaymentRuleBillingScope,
+    createTestPaymentRule,
+} = require('@condo/domains/acquiring/utils/testSchema')
+const { addAcquiringIntegrationAndContext } = require('@condo/domains/acquiring/utils/testSchema')
+const { createTestBankAccount } = require('@condo/domains/banking/utils/testSchema')
+const { createTestOrganization } = require('@condo/domains/organization/utils/testSchema')
+const { createTestProperty } = require('@condo/domains/property/utils/testSchema')
 const { makeClientWithNewRegisteredAndLoggedInUser, makeClientWithSupportUser } = require('@condo/domains/user/utils/testSchema')
 
-const { PaymentRuleBillingScope, createTestPaymentRuleBillingScope, updateTestPaymentRuleBillingScope,
-    createTestPaymentRule, updateTestPaymentRule
-} = require('@condo/domains/acquiring/utils/testSchema')
-const {createTestOrganization} = require("../../organization/utils/testSchema");
-const {addAcquiringIntegrationAndContext, createTestPaymentRuleMarketPlaceScope} = require("../utils/testSchema");
-const {createTestBankAccount} = require("../../banking/utils/testSchema");
-const {createTestProperty} = require("@condo/domains/property/utils/testSchema");
+
 
 function expectedFields (client, paymentRuleBillingScope, paymentRule, attrs) {
     return {
@@ -219,7 +219,7 @@ describe('PaymentRuleBillingScope', () => {
             const [anotherProperty] = await createTestProperty(admin, anotherOrganization)
             await expectToThrowValidationFailureError(async () =>
                 await createTestPaymentRuleBillingScope(admin, paymentRule, { property: { connect: { id: anotherProperty.id } } }),
-                'Property must be of same organization, as payment rule'
+            'Property must be of same organization, as payment rule'
             )
         })
 
@@ -235,16 +235,16 @@ describe('PaymentRuleBillingScope', () => {
                 [['']],
                 [[null]],
                 [[undefined]],
-                [[1,2,3]],
+                [[1, 2, 3]],
             ]
-            test.each(invalidServiceIds)(`Throws validation error for serviceIds: %p`, async (serviceIds) => {
+            test.each(invalidServiceIds)('Throws validation error for serviceIds: %p', async (serviceIds) => {
                 await expectToThrowGQLError(async () =>
-                        await createTestPaymentRuleBillingScope(admin, paymentRule, { serviceIds }),
-                    {
-                        code: 'BAD_USER_INPUT',
-                        type: 'PAYMENT_RULE_BILLING_SCOPE_INVALID_SERVICE_IDS',
-                        message: '"serviceIds" field validation error. JSON was not in the correct format',
-                    }
+                    await createTestPaymentRuleBillingScope(admin, paymentRule, { serviceIds }),
+                {
+                    code: 'BAD_USER_INPUT',
+                    type: 'PAYMENT_RULE_BILLING_SCOPE_INVALID_SERVICE_IDS',
+                    message: '"serviceIds" field validation error. JSON was not in the correct format',
+                }
                 )
             })
         })
