+++ apps/condo/domains/acquiring/schema/PaymentRuleMarketScope.js
@@ -1,0 +1,106 @@
+/**
+ * Generated by `createschema acquiring.PaymentRuleMarketScope 'paymentRule:Relationship:PaymentRule:PROTECT;property?:Relationship:Property:PROTECT;squIds?:Json'`
+ */
+
+const Ajv = require('ajv')
+const { get, pick } = require('lodash')
+
+const { historical, versioned, uuided, tracked, softDeleted, dvAndSender } = require('@open-condo/keystone/plugins')
+const { GQLListSchema, getById } = require('@open-condo/keystone/schema')
+
+const access = require('@condo/domains/acquiring/access/PaymentRuleMarketScope')
+const { getGQLErrorValidator } = require('@condo/domains/common/schema/json.utils')
+
+const { PAYMENT_RULE_MARKET_SCOPE_INVALID_SKU_IDS, PAYMENT_RULE_MARKET_SCOPE_MISSING_ALL_FIELDS, PAYMENT_RULE_MARKET_SCOPE_PROPERTY_AND_PAYMENT_RULE_ORGANIZATIONS_DOES_NOT_MATCH } = require('../constants/errors')
+
+const ajv = new Ajv()
+
+const validateSkuIdsSchema = getGQLErrorValidator(ajv.compile({
+    type: 'array',
+    items: {
+        type: 'string',
+        minLength: 1,
+    },
+}), PAYMENT_RULE_MARKET_SCOPE_INVALID_SKU_IDS)
+
+
+const PaymentRuleMarketScope = new GQLListSchema('PaymentRuleMarketScope', {
+    schemaDoc: 'Conditions to match paymentRule with the marketplace invoices',
+    fields: {
+
+        paymentRule: {
+            schemaDoc: 'Payment rule to apply if market place invoice matches the scope',
+            type: 'Relationship',
+            ref: 'PaymentRule',
+            isRequired: true,
+            knexOptions: { isNotNullable: true }, // Required relationship only!
+            kmigratorOptions: { null: false, on_delete: 'models.PROTECT' },
+        },
+
+        property: {
+            schemaDoc: 'Address restrictions',
+            type: 'Relationship',
+            ref: 'Property',
+            isRequired: false,
+            kmigratorOptions: { null: true, on_delete: 'models.PROTECT' },
+            hooks: {
+                validateInput: async ({ resolvedData, fieldPath, addFieldValidationError }) => {
+                    const propertyId = get(resolvedData, fieldPath)
+                    if (!propertyId) {
+                        return
+                    }
+
+                    const paymentRuleId = get(resolvedData, 'paymentRule')
+                    const property = await getById('Property', propertyId)
+                    const paymentRule = await getById('PaymentRule', paymentRuleId)
+                    if (paymentRule.organization !== property.organization) {
+                        return addFieldValidationError(PAYMENT_RULE_MARKET_SCOPE_PROPERTY_AND_PAYMENT_RULE_ORGANIZATIONS_DOES_NOT_MATCH)
+                    }
+                },
+            },
+        },
+
+        skuIds: {
+            schemaDoc: 'Additional payment split according to the market items in the invoice',
+            type: 'Json',
+            isRequired: false,
+            hooks: {
+                validateInput: ({ resolvedData, fieldPath, context }) => {
+                    const skuIds = get(resolvedData, fieldPath)
+                    if (skuIds === null || skuIds === undefined) {
+                        return
+                    }
+                    validateSkuIdsSchema({ resolvedData, fieldPath, context })
+                },
+            },
+        },
+
+    },
+    hooks: {
+        validateInput: async ({ resolvedData, addValidationError }) => {
+            const notRequiredFields = Object.keys(pick(resolvedData, ['skuIds', 'property']))
+            const notEmptyNotRequiredFields = notRequiredFields.filter(field => !!resolvedData[field])
+            console.error({
+                notRequiredFields,
+                notEmptyNotRequiredFields,
+                resolvedData,
+            })
+            if (!notEmptyNotRequiredFields.length) {
+                return addValidationError(PAYMENT_RULE_MARKET_SCOPE_MISSING_ALL_FIELDS)
+            }
+
+        },
+    },
+    plugins: [uuided(), versioned(), tracked(), softDeleted(), dvAndSender(), historical()],
+    access: {
+        read: access.canReadPaymentRuleMarketScopes,
+        create: access.canManagePaymentRuleMarketScopes,
+        update: access.canManagePaymentRuleMarketScopes,
+        delete: false,
+        auth: true,
+    },
+})
+
+module.exports = {
+    PaymentRuleMarketScope: PaymentRuleMarketScope,
+}
