+++ apps/condo/domains/acquiring/utils/clientSchema/PaymentRule.ts
@@ -1,5 +1,5 @@
 /**
- * Generated by `createschema acquiring.PaymentRule 'context:Relationship:AcquiringIntegrationContext:PROTECT;explicitFee?:Decimal;explicitServiceCharge?:Decimal;implicitFee?:Decimal;bankAccount?:Relationship:BankAccount:PROTECT;marketPlaceScope?:Relationship:PaymentRuleMarketPlaceScope:PROTECT;billingScope?:Relationship:PaymentRuleBillingScope:PROTECT'`
+ * Generated by `createschema acquiring.PaymentRule 'context:Relationship:AcquiringIntegrationContext:PROTECT;explicitFee?:Decimal;explicitServiceCharge?:Decimal;implicitFee?:Decimal;bankAccount?:Relationship:BankAccount:PROTECT;marketPlaceScope?:Relationship:PaymentRuleMarketScope:PROTECT;billingScope?:Relationship:PaymentRuleBillingScope:PROTECT'`
  */
 
 import {
