--- apps/condo/domains/acquiring/utils/clientSchema/PaymentRuleMarketPlaceScope.ts
@@ -1,30 +1,0 @@
-/**
- * Generated by `createschema acquiring.PaymentRuleMarketPlaceScope 'paymentRule:Relationship:PaymentRule:PROTECT;property?:Relationship:Property:PROTECT;squIds?:Json'`
- */
-
-import {
-    PaymentRuleMarketPlaceScope,
-    PaymentRuleMarketPlaceScopeCreateInput,
-    PaymentRuleMarketPlaceScopeUpdateInput,
-    QueryAllPaymentRuleMarketPlaceScopesArgs,
-} from '@app/condo/schema'
-import { generateReactHooks } from '@open-condo/codegen/generate.hooks'
-import { PaymentRuleMarketPlaceScope as PaymentRuleMarketPlaceScopeGQL } from '@condo/domains/acquiring/gql'
-
-// TODO(codegen): write utils like convertToFormState and formValuesProcessor if needed, otherwise delete this TODO
-
-const {
-    useObject,
-    useObjects,
-    useCreate,
-    useUpdate,
-    useSoftDelete,
-} = generateReactHooks<PaymentRuleMarketPlaceScope, PaymentRuleMarketPlaceScopeCreateInput, PaymentRuleMarketPlaceScopeUpdateInput, QueryAllPaymentRuleMarketPlaceScopesArgs>(PaymentRuleMarketPlaceScopeGQL)
-
-export {
-    useObject,
-    useObjects,
-    useCreate,
-    useUpdate,
-    useSoftDelete,
-}
