+++ apps/condo/domains/acquiring/utils/clientSchema/PaymentRuleMarketScope.ts
@@ -1,0 +1,30 @@
+/**
+ * Generated by `createschema acquiring.PaymentRuleMarketScope 'paymentRule:Relationship:PaymentRule:PROTECT;property?:Relationship:Property:PROTECT;squIds?:Json'`
+ */
+
+import {
+    PaymentRuleMarketScope,
+    PaymentRuleMarketScopeCreateInput,
+    PaymentRuleMarketScopeUpdateInput,
+    QueryAllPaymentRuleMarketScopesArgs,
+} from '@app/condo/schema'
+
+import { generateReactHooks } from '@open-condo/codegen/generate.hooks'
+
+import { PaymentRuleMarketScope as PaymentRuleMarketScopeGQL } from '@condo/domains/acquiring/gql'
+
+const {
+    useObject,
+    useObjects,
+    useCreate,
+    useUpdate,
+    useSoftDelete,
+} = generateReactHooks<PaymentRuleMarketScope, PaymentRuleMarketScopeCreateInput, PaymentRuleMarketScopeUpdateInput, QueryAllPaymentRuleMarketScopesArgs>(PaymentRuleMarketScopeGQL)
+
+export {
+    useObject,
+    useObjects,
+    useCreate,
+    useUpdate,
+    useSoftDelete,
+}
