+++ apps/condo/domains/acquiring/utils/clientSchema/index.ts
@@ -11,5 +11,5 @@ export * as Payment from './Payment'
 export * as PaymentsFilterTemplate from './PaymentsFilterTemplate'
 export * as PaymentRule from './PaymentRule'
 export * as PaymentRuleBillingScope from './PaymentRuleBillingScope'
-export * as PaymentRuleMarketPlaceScope from './PaymentRuleMarketPlaceScope'
+export * as PaymentRuleMarketScope from './PaymentRuleMarketScope'
 /* AUTOGENERATE MARKER <IMPORT-EXPORT> */
