+++ apps/condo/domains/acquiring/utils/testSchema/index.js
@@ -57,7 +57,7 @@ const { DEFAULT_ORGANIZATION_TIMEZONE } = require('@condo/domains/organization/c
 const { CALCULATE_FEE_FOR_RECEIPT_QUERY } = require('@condo/domains/acquiring/gql')
 const { PaymentRule: PaymentRuleGQL } = require('@condo/domains/acquiring/gql')
 const { PaymentRuleBillingScope: PaymentRuleBillingScopeGQL } = require('@condo/domains/acquiring/gql')
-const { PaymentRuleMarketPlaceScope: PaymentRuleMarketPlaceScopeGQL } = require('@condo/domains/acquiring/gql')
+const { PaymentRuleMarketScope: PaymentRuleMarketScopeGQL } = require('@condo/domains/acquiring/gql')
 /* AUTOGENERATE MARKER <IMPORT> */
 
 const AcquiringIntegration = generateGQLTestUtils(AcquiringIntegrationGQL)
@@ -70,7 +70,7 @@ const RecurrentPaymentContext = generateGQLTestUtils(RecurrentPaymentContextGQL)
 const RecurrentPayment = generateGQLTestUtils(RecurrentPaymentGQL)
 const PaymentRule = generateGQLTestUtils(PaymentRuleGQL)
 const PaymentRuleBillingScope = generateGQLTestUtils(PaymentRuleBillingScopeGQL)
-const PaymentRuleMarketPlaceScope = generateGQLTestUtils(PaymentRuleMarketPlaceScopeGQL)
+const PaymentRuleMarketScope = generateGQLTestUtils(PaymentRuleMarketScopeGQL)
 /* AUTOGENERATE MARKER <CONST> */
 
 const RecurrentPaymentContextLiteGQL = generateGqlQueries('RecurrentPaymentContext', '{ id }')
@@ -608,12 +608,10 @@ async function updateTestPaymentRuleBillingScope (client, id, extraAttrs = {}) {
     return [obj, attrs]
 }
 
-async function createTestPaymentRuleMarketPlaceScope (client, paymentRule, extraAttrs = {}) {
+async function createTestPaymentRuleMarketScope (client, paymentRule, extraAttrs = {}) {
     if (!client) throw new Error('no client')
     if (!paymentRule || !paymentRule.id) throw new Error('no paymentRule.id')
     const sender = { dv: 1, fingerprint: faker.random.alphaNumeric(8) }
-
-    // TODO(codegen): write createTestPaymentRuleMarketPlaceScope logic for generate fields
 
     const attrs = {
         dv: 1,
@@ -621,23 +619,21 @@ async function createTestPaymentRuleMarketPlaceScope (client, paymentRule, extra
         paymentRule: { connect: { id: paymentRule.id } },
         ...extraAttrs,
     }
-    const obj = await PaymentRuleMarketPlaceScope.create(client, attrs)
+    const obj = await PaymentRuleMarketScope.create(client, attrs)
     return [obj, attrs]
 }
 
-async function updateTestPaymentRuleMarketPlaceScope (client, id, extraAttrs = {}) {
+async function updateTestPaymentRuleMarketScope (client, id, extraAttrs = {}) {
     if (!client) throw new Error('no client')
     if (!id) throw new Error('no id')
     const sender = { dv: 1, fingerprint: faker.random.alphaNumeric(8) }
-
-    // TODO(codegen): check the updateTestPaymentRuleMarketPlaceScope logic for generate fields
 
     const attrs = {
         dv: 1,
         sender,
         ...extraAttrs,
     }
-    const obj = await PaymentRuleMarketPlaceScope.update(client, id, attrs)
+    const obj = await PaymentRuleMarketScope.update(client, id, attrs)
     return [obj, attrs]
 }
 
@@ -913,6 +909,6 @@ module.exports = {
     calculateFeeForReceiptByTestClient,
     PaymentRule, createTestPaymentRule, updateTestPaymentRule,
     PaymentRuleBillingScope, createTestPaymentRuleBillingScope, updateTestPaymentRuleBillingScope,
-    PaymentRuleMarketPlaceScope, createTestPaymentRuleMarketPlaceScope, updateTestPaymentRuleMarketPlaceScope,
+    PaymentRuleMarketScope, createTestPaymentRuleMarketScope, updateTestPaymentRuleMarketScope,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
